Demo/requirements_demo.txt
streamlit>=1.32.0
plotly>=5.20.0
torch>=2.0.0
numpy>=1.24.0
pandas>=1.5.0
pyyaml>=6.0
openpyxl>=3.0.0
